% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.nvar.r
\name{calc.nvar}
\alias{calc.nvar}
\title{Calculate NVAR}
\usage{
calc.nvar(rough = NA, DEM = NA, x, resamp = FALSE)
}
\arguments{
\item{rough}{SpatRaster object containing roughness information. Might be NA. See details.}

\item{DEM}{SpatRaster object containing a digital elevation model where roughness information can be derived from. Might be NA. See details.}

\item{x}{SpatVector object with points geometry. At these locations, NVAR will be extracted.}

\item{resamp}{SpatRaster object or FALSE. If resamp is a SpatRaster object, rough is resampled to match the spatial resolution of resamp.}
}
\value{
A two column matrix containing the ID of the points in x and the computed values of NVAR.
}
\description{
Calculate COSERO's parameter NVAR from roughness properties derived from a DEM
}
\details{
This function replaced \\link{compute.nvar}, which is deprecated. It calculates NVAR at certain points given by the parameter x. Eigher DEM or rough must be given, the other one then need's to be NA. If the first, roughness will be calculated.
}
\author{
Simon Frey
}
