% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmplt.qobs.r
\name{cmplt.qobs}
\alias{cmplt.qobs}
\title{Complete QOBS for use in COSERO}
\usage{
cmplt.qobs(x, start = NULL, end = NULL, TZ = "utc", file = NULL,
  ...)
}
\arguments{
\item{x}{an xts object}

\item{start}{a POSIXct object or a character string that can be coerced to such. Gives the desired start time of the time series.}

\item{end}{a POSIXct object or a character string that can be coerced to such. Gives the desired end time of the time series.}

\item{TZ}{time zone specification to be used for the conversion, if one is required. System-specific (see time zones), but "" is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning.}

\item{file}{name of the outputfile or NULL. If the latter, no file will be written.}

\item{...}{additional parameters passed on to \code{\link{as.POSIXct}}}
}
\value{
an xts object
}
\description{
Stick together an xts time series with start and end dates that might lay outside the original time series
}
\details{
An xts object is extended at the beginning and end of the dates. Duplicated values (as they might exist because of time changes) are removed. NA values are substitued with -0.01 which is COSEROS standard for missing.

   If file != NULL a txt file will be written using \code{\link{write.xts}} from the TigR package
}
\author{
Simon Frey
}
