% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tmmon.r
\name{compute_tmmon}
\alias{compute_tmmon}
\title{Compute tmmon from a T_IZ file for COSERO}
\usage{
compute_tmmon(x, filename = NULL, ...)
}
\arguments{
\item{x}{an xts object}

\item{filename}{NULL or a character string representing the filename of the outputfile}

\item{...}{further arguments passed on to \code{\link{apply.monthly}}}
}
\value{
a vector or nothing if filename != NULL
}
\description{
Compute the parameter tmmon (mean monthly temperature for use in the Thornthwaite method) from a T_IZ file
}
\author{
Simon Frey
}
