% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nvar.r
\name{compute_nvar}
\alias{compute_nvar}
\title{Compute the nvar parameter from a DEM}
\usage{
compute_nvar(SP, DEM, NZ)
}
\arguments{
\item{SP}{a spatialPoints object at which points the values of the raster should be extracted or a character string pointing at a ESRI shapefile}

\item{DEM}{a raster object containing the DEM information or a character string pointing at such a raster}

\item{NZ}{a raster object containing the information of the NZ values of the COSERO model or a character string pointing at such a raster}
}
\value{
Returns a matrix of nvar values at specific points
}
\description{
Derive the parameter nvar for the hydrological model COSERO from roughness parameters of a DEM
}
\author{
Simon Frey
}
